Improved hlsl shader code readability Optimized rendering loop Separated shape instances that are declared once and immediate shape instances that are declared each frame. im_instances and instances. Created vec2i for easier use of functions.

+++ flashgui/dxgicontext.h
@@ -83,7 +83,7 @@ namespace fgui {
 				return {};
 
 			D3D12_CPU_DESCRIPTOR_HANDLE handle = rtv_heap->GetCPUDescriptorHandleForHeapStart();
-			handle.ptr += index * rtv_descriptor_size;
+			handle.ptr += static_cast<SIZE_T>(index) * rtv_descriptor_size;
 
 			return handle;
 		}
